modified:   client/src/App.js modified:   client/src/components/Login.js     fixed import of provider since I change the Modal file structure
modified:   client/src/components/Modal/components/Modal.js
    Removed button and link to its own file

new file:   client/src/components/Modal/components/ModalButton.js
    Created new file for Modal Button
new file:   client/src/components/Modal/components/ModalLink.js
    Created new file for ModalLink
modified:   client/src/components/Modal/index.js
    index file exports all Modal components
renamed:    client/src/components/Modal/modalStore/modalActions.js -> client/src/components/Modal/modalActions/index.js
    moved modalAction in its own folder
modified:   client/src/components/Modal/modalStore/index.js
    moved the modal context to its own folder
modified:   client/src/components/Modal/modalStore/modalReducer.js
    moved reducer to modalStore folder, fixed imports
modified:   client/src/layouts/MainWithLogin.js
    removed unnecessary imports
modified:   client/src/utils/modalHooks.js
    fixed imports
