+++ client/src/App.js
@@ -10,7 +10,7 @@ import loadGlobalIcons from "./utils/icons";
 import { useSocketConnection } from "./utils/socket.io";
 import Fade from "./animations/Fade";
 import { useUserSocketUpdates } from "utils/user";
-import { ModalProvider } from "components/Modal/components/Modal";
+import { ModalProvider } from "components/Modal/modalStore";
 
 loadGlobalIcons();
