+++ client/src/components/Login.js
@@ -8,8 +8,8 @@ import { useLogout, useLogin } from "utils/auth";
 import Form from "./Form";
 import { createValidator } from "utils/validation";
 import { useHistory } from "react-router-dom";
-import { ModalButton, ModalLink } from "./Modal/components/Modal";
-import { Modal } from "./Modal";
+import { ModalButton, ModalLink, Modal } from "./Modal";
+
 
 const validateLoginData = createValidator({
     validators: {
