+++ client/src/components/Modal/components/Modal.js
@@ -1,91 +1,11 @@
-import { useContext, createContext, useReducer } from "react";
-
 import {
     Modal as BulmaModal,
-    Button
 } from "react-bulma-components";
-import { SET_ACTIVE_MODAL } from "./modalActions";
+import { SET_ACTIVE_MODAL } from "../modalActions";
+import { useModalContext } from "../modalStore";
 
-import reducer from "./modalReducer";
 
-// Define a new context
-const ModalContext = createContext(false);
-
-// Deconstruct the provider for ease of use in JSX
-const { Provider } = ModalContext;
-
-export const useModalContext = () => {
-
-    return useContext( ModalContext );
-    
-}
-
-/**
- * Modal provider component.
- * @param {object} props 
- */
-export const ModalProvider = ({children, isActive = false}) => {
-
-    // Create the reducer state.
-     const [modalState, modalDispatch] = useReducer(reducer, { 
-            modals: { } , 
-            activeKey: "" 
-        } )
-    
-    //const modalState = useState( isActive );
-
-    return <Provider value={[modalState, modalDispatch]}>{children}</Provider>;
-
-}
-
-/**
- * Open modal button component. Requires <ModalProvider> as an ancenstor.
- * @param {object} props
- */
-export const ModalButton = ({ children, modalKey, ...props }) => {
-	// Consume the login context to fetch the live state.
-    const [, modalDispatch] = useModalContext();
-
-    const open = () =>
-		modalDispatch({
-			type: SET_ACTIVE_MODAL,
-			payload: modalKey
-		});
-
-	return (
-		<Button onClick={() => open()} {...props}>
-			{children || "Launch Modal"}
-		</Button>
-	);
-};
-
-/**
- * Open modal link component. Requires <ModalProvider> as an ancenstor.
- * @param {object} props
- */
-export const ModalLink = ({ children, onClick, modalKey, ...props }) => {
-	// Consume the login context to fetch the live state.
-
-	const [, modalDispatch] = useModalContext();
-
-	const open = () =>
-		modalDispatch({
-			type: SET_ACTIVE_MODAL,
-			payload: modalKey
-		});
-
-	const onLinkClick = () => {
-		onClick ? onClick(open) : open();
-	};
-
-	return (
-		<a role="button" onClick={onLinkClick} {...props}>
-			{children || "Launch Modal"}
-		</a>
-	);
-};;
-
-const Modal = ( { children, onClose, contentProps = {}, ...props } ) => {
+export const Modal = ( { children, onClose, contentProps = {}, ...props } ) => {
 
 
     const [modal, modalDispatch] = useModalContext();
@@ -111,5 +31,3 @@ const Modal = ( { children, onClose, contentProps = {}, ...props } ) => {
     )
 
 }
-
-export default Modal;
