+++ client/src/components/Modal/index.js
@@ -1,1 +1,3 @@
 export { Modal } from "./components/Modal";
+export { ModalButton } from "./components/ModalButton";
+export { ModalLink } from "./components/ModalLink";
