+++ client/src/layouts/MainWithLogin.js
@@ -1,6 +1,5 @@
 import Main from "./Main";
 import { LoginModal } from "components/Login"
-import { ModalProvider } from "components/Modal/Modal";
 
 const MainWithLogin = ({ children }) => {
     return (
